syslogformat: move timestamp parsing code to a separate .c file
This patch starts extracting the timestamp parsing code to a separate .c
file.  Since the coupling is not completely removed, this is now included
into the syslog-format.c file, but eventually in a followup patch, this
would become a proper c module.
